Update HashSetStore data store type

+++ src/main/kotlin/pw/zoomer/ipcounter/data/HashSetStore.kt
@@ -3,17 +3,17 @@ package pw.zoomer.ipcounter.data
 import pw.zoomer.ipcounter.net.IpV4
 
 class HashSetStore : IpStore {
-    private val highIPs = HashSet<UInt>()
-    private val lowIPs = HashSet<UInt>()
+    private val highIPs = HashSet<Long>()
+    private val lowIPs = HashSet<Long>()
 
     override val count: Long
         get() = highIPs.size.toLong() + lowIPs.size
 
     override fun add(ipAddress: IpV4) {
         if (ipAddress.octet1 and 0b1000_0000 == 0) {
-            lowIPs.add(ipAddress.toUInt())
+            lowIPs.add(ipAddress.toLong())
         } else {
-            highIPs.add(ipAddress.toUInt())
+            highIPs.add(ipAddress.toLong())
         }
     }
 }
